backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.4
rq==1.16.2
boto3==1.34.107
httpx==0.27.0
python-dotenv==1.0.1
pydantic>=2.9.2
pydantic-settings>=2.2.1
# Optional: opengradient SDK when available
# opengradient>=0.1.0
opengradient
